requirements.txt
astroid==2.1.0
awscli==1.16.81
botocore==1.12.71
colorama==0.3.9
couchbase==2.5.2
docutils==0.14
Faker==1.0.1
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pyasn1==0.4.5
pylint==2.2.2
python-dateutil==2.7.5
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
text-unidecode==1.2
typing==3.6.6
urllib3==1.24.1
wrapt==1.10.11
